fix: Moved LogViewHolder to a separate class for better modularity.
